Split list and iterator from index
